refactor: Export "EditorProps" from editor/index
