+++ src/editor/index.ts
@@ -1,1 +1,2 @@
+export { EditorProps } from './editor'
 export { createPlugins } from './plugins'
